initial steps for the refuge web ui.
The refuge UI will be launched by default on the port 9000 and
advertised over DNSD. For now /local give access to the local couchdb
store and / display a welcome.

With this changes, https is the only publicly advertized transport over
UPNP and dnsd and will be use to transfert data between node. The port
is now auto detected and will change on startup.